Moved semaphore implementations into platform specific files
